ed-devtracker-collector: Fixup for Frontier-QA duplicating a post from long ago.
This is
https://forums.frontier.co.uk/showthread.php/290119-Bug-Reporting-Guidelines-and-Template
which still has a date of 2015-03-02 on it in the ActivityList, so our
"find last post by this user" check doesn't pick it up.

This change appears to duplicate some work in regexes, but that's so
that we don't get an error trying to reference $+{'end'} when it isn't
present in the second form of each regex.