Changes scraper to only ignore 'seen that' posts individually, and not assume that the first one it sees means that it's up to date.
Specifically we just had a post by BrettC with 2.2.03 beta patch notes
not show up.  We *think* this was because he'd started the post in some
hidden state, posted something else, and then un-hidden the patch notes
post.  So the scraper saw the later post and bailed without reaching the
patch notes post it hadn't already seen.