Implements 'UPSERT' for posts, we can do it properly once on Postgresql 9.4. This is the shonky 'SELECT then INSERT/UPPDATE' as appropiate method, with the race condition, but that shouldn't affect *this* use.
This will have the side effect of UPDATE'ing some posts on *every* run.
They're showing up in "Frontier QA"'s Activity feed due to a post being
moved/edited for when 2.3 beta starts, but it's an old, old, post from
2015-03-02 for whatever beta started then.  There's no edit date on it
at all (let alone on the Activity listing), so it'll continue to be one
of the latest posts by that account.