Adds watchdog.sh to check for hung processes.
Despite trying to tell the perl HTTP code to timeout after 10s we were
seeing processes hang waiting on network data for a long time (30+
minutes).  So if such issues arise run this.

It could be improved to specifically checkout lsof output and only count
up if it's seeing the same connection sticking.