First pass of refactor adding ED/DevTracker/Scrape.pm
The actual "retrieve a web page and grab data" parts of the main script
are now in this new file.  This will allow re-use in things like the
pending backfill-fulltext script, or even an artificial post injection
script.

This all appears to be working, but does need testing in detail.  Still
have debug output active.