Some tweaks to backfill script to ensure proper exit, and chance to catch errors.
Obviously we can't continue 'under' database row id of 1.

And we kept on hitting '400 - bad request' errors. Once one was
encountered all subsequent requests also hit that error.  Restarting the
script, with $lastid appropriately initialised, caused the same queries
to succeed.  So there's a sleep there in order to catch this happening.
  Yes, ideally we'd pass the specific error back up from
$scrape->get_fulltext(), but that requires also changing the main
scraper so will have to wait.