Implements forum ignoring.
  This is based on the breadcrumbs when we "click through" to gather
full post text.  It takes the last 'navbit' (only) class one, which
should be the sub-forum the post is in, then uses the *url*, not name,
of this to compare to a list of ignored forums.

  If the forum is ever re-arranged such that another forum takes on the
same number as an ignored one then that 'new' forum is going to be
accidentally ignored.  Else no new forums will be ignored by this.